app/requirements.txt
flask 
pandas
SQLAlchemy
python-dotenv
Flask-Session
gunicorn
sqlalchemy
authlib
requests
flask-sqlalchemy